Crate nserror

Constants

NS_BASE_STREAM_BAD_CONVERSION

NS_BASE_STREAM_CLOSED

NS_BASE_STREAM_ILLEGAL_ARGS

NS_BASE_STREAM_NO_CONVERTER

NS_BASE_STREAM_OSERROR

NS_BASE_STREAM_WOULD_BLOCK

NS_BINDING_ABORTED

NS_BINDING_FAILED

NS_BINDING_REDIRECTED

NS_BINDING_RETARGETED

NS_BINDING_SUCCEEDED

NS_CONTENT_BLOCKED

NS_CONTENT_BLOCKED_SHOW_ALT

NS_ERROR_ABORT

NS_ERROR_ALREADY_CONNECTED

NS_ERROR_ALREADY_INITIALIZED

NS_ERROR_ALREADY_OPENED

NS_ERROR_BASE

NS_ERROR_BLOCKED_URI

NS_ERROR_BRIDGE_OPEN_CHILD

NS_ERROR_BRIDGE_OPEN_PARENT

NS_ERROR_CACHE_DATA_IS_NOT_STREAM

NS_ERROR_CACHE_DATA_IS_STREAM

NS_ERROR_CACHE_ENTRY_DOOMED

NS_ERROR_CACHE_IN_USE

NS_ERROR_CACHE_KEY_NOT_FOUND

NS_ERROR_CACHE_READ_ACCESS_DENIED

NS_ERROR_CACHE_WAIT_FOR_VALIDATION

NS_ERROR_CACHE_WRITE_ACCESS_DENIED

NS_ERROR_CANNOT_CONVERT_DATA

NS_ERROR_CMS_ENCRYPT_INCOMPLETE

NS_ERROR_CMS_ENCRYPT_NO_BULK_ALG

NS_ERROR_CMS_VERIFY_BAD_DIGEST

NS_ERROR_CMS_VERIFY_BAD_SIGNATURE

NS_ERROR_CMS_VERIFY_CERT_WITHOUT_ADDRESS

NS_ERROR_CMS_VERIFY_DIGEST_MISMATCH

NS_ERROR_CMS_VERIFY_ERROR_PROCESSING

NS_ERROR_CMS_VERIFY_ERROR_UNVERIFIED

NS_ERROR_CMS_VERIFY_HEADER_MISMATCH

NS_ERROR_CMS_VERIFY_MALFORMED_SIGNATURE

NS_ERROR_CMS_VERIFY_NOCERT

NS_ERROR_CMS_VERIFY_NOT_SIGNED

NS_ERROR_CMS_VERIFY_NOT_YET_ATTEMPTED

NS_ERROR_CMS_VERIFY_NO_CONTENT_INFO

NS_ERROR_CMS_VERIFY_UNKNOWN_ALGO

NS_ERROR_CMS_VERIFY_UNSUPPORTED_ALGO

NS_ERROR_CMS_VERIFY_UNTRUSTED

NS_ERROR_CONNECTION_REFUSED

NS_ERROR_CONTENT_BLOCKED

NS_ERROR_CONTENT_BLOCKED_SHOW_ALT

NS_ERROR_CONTENT_CRASHED

NS_ERROR_CORRUPTED_CONTENT

NS_ERROR_CSP_FORM_ACTION_VIOLATION

NS_ERROR_CSP_FRAME_ANCESTOR_VIOLATION

NS_ERROR_DNS_LOOKUP_QUEUE_FULL

NS_ERROR_DOCSHELL_DYING

NS_ERROR_DOCSHELL_REQUEST_REJECTED

NS_ERROR_DOCUMENT_IS_PRINTMODE

NS_ERROR_DOCUMENT_NOT_CACHED

NS_ERROR_DOM_ABORT_ERR

NS_ERROR_DOM_ANIM_MISSING_PROPS_ERR

NS_ERROR_DOM_BAD_DOCUMENT_DOMAIN

NS_ERROR_DOM_BAD_URI

NS_ERROR_DOM_DATA_CLONE_ERR

NS_ERROR_DOM_DATA_ERR

NS_ERROR_DOM_ENCODING_NOT_SUPPORTED_ERR

NS_ERROR_DOM_FILEHANDLE_ABORT_ERR

NS_ERROR_DOM_FILEHANDLE_INACTIVE_ERR

NS_ERROR_DOM_FILEHANDLE_NOT_ALLOWED_ERR

NS_ERROR_DOM_FILEHANDLE_QUOTA_ERR

NS_ERROR_DOM_FILEHANDLE_READ_ONLY_ERR

NS_ERROR_DOM_FILEHANDLE_UNKNOWN_ERR

NS_ERROR_DOM_FILESYSTEM_INVALID_MODIFICATION_ERR

NS_ERROR_DOM_FILESYSTEM_INVALID_PATH_ERR

NS_ERROR_DOM_FILESYSTEM_NO_MODIFICATION_ALLOWED_ERR

NS_ERROR_DOM_FILESYSTEM_PATH_EXISTS_ERR

NS_ERROR_DOM_FILESYSTEM_TYPE_MISMATCH_ERR

NS_ERROR_DOM_FILESYSTEM_UNKNOWN_ERR

NS_ERROR_DOM_FILE_ABORT_ERR

NS_ERROR_DOM_FILE_NOT_FOUND_ERR

NS_ERROR_DOM_FILE_NOT_READABLE_ERR

NS_ERROR_DOM_HIERARCHY_REQUEST_ERR

NS_ERROR_DOM_INDEXEDDB_ABORT_ERR

NS_ERROR_DOM_INDEXEDDB_CONSTRAINT_ERR

NS_ERROR_DOM_INDEXEDDB_DATA_ERR

NS_ERROR_DOM_INDEXEDDB_NOT_ALLOWED_ERR

NS_ERROR_DOM_INDEXEDDB_NOT_FOUND_ERR

NS_ERROR_DOM_INDEXEDDB_QUOTA_ERR

NS_ERROR_DOM_INDEXEDDB_READ_ONLY_ERR

NS_ERROR_DOM_INDEXEDDB_RECOVERABLE_ERR

NS_ERROR_DOM_INDEXEDDB_TIMEOUT_ERR

NS_ERROR_DOM_INDEXEDDB_TRANSACTION_INACTIVE_ERR

NS_ERROR_DOM_INDEXEDDB_UNKNOWN_ERR

NS_ERROR_DOM_INDEXEDDB_VERSION_ERR

NS_ERROR_DOM_INDEX_SIZE_ERR

NS_ERROR_DOM_INUSE_ATTRIBUTE_ERR

NS_ERROR_DOM_INVALID_ACCESS_ERR

NS_ERROR_DOM_INVALID_ACCESS_XHR_TIMEOUT_AND_RESPONSETYPE_UNSUPPORTED_FOR_SYNC

NS_ERROR_DOM_INVALID_CHARACTER_ERR

NS_ERROR_DOM_INVALID_EXPRESSION_ERR

NS_ERROR_DOM_INVALID_HEADER_NAME

NS_ERROR_DOM_INVALID_MODIFICATION_ERR

NS_ERROR_DOM_INVALID_NODE_TYPE_ERR

NS_ERROR_DOM_INVALID_POINTER_ERR

NS_ERROR_DOM_INVALID_STATE_ERR

NS_ERROR_DOM_INVALID_STATE_XHR_CHUNKED_RESPONSETYPES_UNSUPPORTED_FOR_SYNC

NS_ERROR_DOM_INVALID_STATE_XHR_HAS_INVALID_CONTEXT

NS_ERROR_DOM_INVALID_STATE_XHR_HAS_WRONG_RESPONSETYPE_FOR_RESPONSETEXT

NS_ERROR_DOM_INVALID_STATE_XHR_HAS_WRONG_RESPONSETYPE_FOR_RESPONSEXML

NS_ERROR_DOM_INVALID_STATE_XHR_MUST_BE_OPENED

NS_ERROR_DOM_INVALID_STATE_XHR_MUST_NOT_BE_LOADING_OR_DONE

NS_ERROR_DOM_INVALID_STATE_XHR_MUST_NOT_BE_SENDING

NS_ERROR_DOM_JS_DECODING_ERROR

NS_ERROR_DOM_MALFORMED_URI

NS_ERROR_DOM_MEDIA_ABORT_ERR

NS_ERROR_DOM_MEDIA_CANCELED

NS_ERROR_DOM_MEDIA_CDM_ERR

NS_ERROR_DOM_MEDIA_CUBEB_INITIALIZATION_ERR

NS_ERROR_DOM_MEDIA_DECODE_ERR

NS_ERROR_DOM_MEDIA_DEMUXER_ERR

NS_ERROR_DOM_MEDIA_END_OF_STREAM

NS_ERROR_DOM_MEDIA_FATAL_ERR

NS_ERROR_DOM_MEDIA_INITIALIZING_DECODER

NS_ERROR_DOM_MEDIA_MEDIASINK_ERR

NS_ERROR_DOM_MEDIA_METADATA_ERR

NS_ERROR_DOM_MEDIA_NEED_NEW_DECODER

NS_ERROR_DOM_MEDIA_NOT_ALLOWED_ERR

NS_ERROR_DOM_MEDIA_NOT_SUPPORTED_ERR

NS_ERROR_DOM_MEDIA_OVERFLOW_ERR

NS_ERROR_DOM_MEDIA_WAITING_FOR_DATA

NS_ERROR_DOM_NAMESPACE_ERR

NS_ERROR_DOM_NETWORK_ERR

NS_ERROR_DOM_NOT_ALLOWED_ERR

NS_ERROR_DOM_NOT_BOOLEAN_ERR

NS_ERROR_DOM_NOT_FOUND_ERR

NS_ERROR_DOM_NOT_FUNCTION_ERR

NS_ERROR_DOM_NOT_NUMBER_ERR

NS_ERROR_DOM_NOT_OBJECT_ERR

NS_ERROR_DOM_NOT_SUPPORTED_ERR

NS_ERROR_DOM_NOT_XPC_OBJECT_ERR

NS_ERROR_DOM_NO_MODIFICATION_ALLOWED_ERR

NS_ERROR_DOM_OPERATION_ERR

NS_ERROR_DOM_PROP_ACCESS_DENIED

NS_ERROR_DOM_PUSH_ABORT_ERR

NS_ERROR_DOM_PUSH_DENIED_ERR

NS_ERROR_DOM_PUSH_GCM_DISABLED

NS_ERROR_DOM_PUSH_INVALID_KEY_ERR

NS_ERROR_DOM_PUSH_INVALID_REGISTRATION_ERR

NS_ERROR_DOM_PUSH_MISMATCHED_KEY_ERR

NS_ERROR_DOM_PUSH_SERVICE_UNREACHABLE

NS_ERROR_DOM_QUOTA_EXCEEDED_ERR

NS_ERROR_DOM_QUOTA_REACHED

NS_ERROR_DOM_RETVAL_UNDEFINED

NS_ERROR_DOM_SECMAN_ERR

NS_ERROR_DOM_SECURITY_ERR

NS_ERROR_DOM_SVG_MATRIX_NOT_INVERTABLE

NS_ERROR_DOM_SVG_WRONG_TYPE_ERR

NS_ERROR_DOM_SYNTAX_ERR

NS_ERROR_DOM_TIMEOUT_ERR

NS_ERROR_DOM_TOO_FEW_PARAMETERS_ERR

NS_ERROR_DOM_TYPE_ERR

NS_ERROR_DOM_TYPE_MISMATCH_ERR

NS_ERROR_DOM_UNKNOWN_ERR

NS_ERROR_DOM_URL_MISMATCH_ERR

NS_ERROR_DOM_WRONG_DOCUMENT_ERR

NS_ERROR_DOM_WRONG_TYPE_ERR

NS_ERROR_DOM_XPCONNECT_ACCESS_DENIED

NS_ERROR_DOWNLOAD_COMPLETE

NS_ERROR_DOWNLOAD_NOT_PARTIAL

NS_ERROR_DUPLICATE_HANDLE

NS_ERROR_ENTITY_CHANGED

NS_ERROR_FACTORY_EXISTS

NS_ERROR_FACTORY_NOT_LOADED

NS_ERROR_FACTORY_NOT_REGISTERED

NS_ERROR_FACTORY_NO_SIGNATURE_SUPPORT

NS_ERROR_FACTORY_REGISTER_AGAIN

NS_ERROR_FAILURE

NS_ERROR_FILE_ACCESS_DENIED

NS_ERROR_FILE_ALREADY_EXISTS

NS_ERROR_FILE_COPY_OR_MOVE_FAILED

NS_ERROR_FILE_CORRUPTED

NS_ERROR_FILE_DESTINATION_NOT_DIR

NS_ERROR_FILE_DIR_NOT_EMPTY

NS_ERROR_FILE_DISK_FULL

NS_ERROR_FILE_EXECUTION_FAILED

NS_ERROR_FILE_INVALID_PATH

NS_ERROR_FILE_IS_DIRECTORY

NS_ERROR_FILE_IS_LOCKED

NS_ERROR_FILE_NAME_TOO_LONG

NS_ERROR_FILE_NOT_DIRECTORY

NS_ERROR_FILE_NOT_FOUND

NS_ERROR_FILE_NO_DEVICE_SPACE

NS_ERROR_FILE_READ_ONLY

NS_ERROR_FILE_TARGET_DOES_NOT_EXIST

NS_ERROR_FILE_TOO_BIG

NS_ERROR_FILE_UNKNOWN_TYPE

NS_ERROR_FILE_UNRECOGNIZED_PATH

NS_ERROR_FILE_UNRESOLVABLE_SYMLINK

NS_ERROR_FIRST_HEADER_FIELD_COMPONENT_EMPTY

NS_ERROR_FTP_CWD

NS_ERROR_FTP_LIST

NS_ERROR_FTP_LOGIN

NS_ERROR_FTP_PASV

NS_ERROR_FTP_PWD

NS_ERROR_GFX_CMAP_MALFORMED

NS_ERROR_GFX_PRINTER_COULD_NOT_OPEN_FILE

NS_ERROR_GFX_PRINTER_DOC_IS_BUSY

NS_ERROR_GFX_PRINTER_ENDDOC

NS_ERROR_GFX_PRINTER_NAME_NOT_FOUND

NS_ERROR_GFX_PRINTER_NO_PRINTER_AVAILABLE

NS_ERROR_GFX_PRINTER_STARTDOC

NS_ERROR_GFX_PRINTER_STARTPAGE

NS_ERROR_HARMFUL_URI

NS_ERROR_HOST_IS_IP_ADDRESS

NS_ERROR_HSTS_PRIMING_TIMEOUT

NS_ERROR_HTMLPARSER_BADATTRIBUTE

NS_ERROR_HTMLPARSER_BADCONTEXT

NS_ERROR_HTMLPARSER_BADDTD

NS_ERROR_HTMLPARSER_BADFILENAME

NS_ERROR_HTMLPARSER_BADTOKENIZER

NS_ERROR_HTMLPARSER_BADURL

NS_ERROR_HTMLPARSER_BLOCK

NS_ERROR_HTMLPARSER_CANTPROPAGATE

NS_ERROR_HTMLPARSER_CONTEXTMISMATCH

NS_ERROR_HTMLPARSER_CONTINUE

NS_ERROR_HTMLPARSER_EOF

NS_ERROR_HTMLPARSER_FAKE_ENDTAG

NS_ERROR_HTMLPARSER_HIERARCHYTOODEEP

NS_ERROR_HTMLPARSER_INTERRUPTED

NS_ERROR_HTMLPARSER_INVALIDPARSERCONTEXT

NS_ERROR_HTMLPARSER_INVALID_COMMENT

NS_ERROR_HTMLPARSER_MISPLACEDTABLECONTENT

NS_ERROR_HTMLPARSER_STOPPARSING

NS_ERROR_HTMLPARSER_UNKNOWN

NS_ERROR_HTMLPARSER_UNRESOLVEDDTD

NS_ERROR_HTMLPARSER_UNTERMINATEDSTRINGLITERAL

NS_ERROR_ILLEGAL_DURING_SHUTDOWN

NS_ERROR_ILLEGAL_INPUT

NS_ERROR_ILLEGAL_VALUE

NS_ERROR_INSUFFICIENT_DOMAIN_LEVELS

NS_ERROR_INTERCEPTION_FAILED

NS_ERROR_INTERNAL_ERRORRESULT_DOMEXCEPTION

NS_ERROR_INTERNAL_ERRORRESULT_EXCEPTION_ON_JSCONTEXT

NS_ERROR_INTERNAL_ERRORRESULT_JS_EXCEPTION

NS_ERROR_INTERNAL_ERRORRESULT_RANGEERROR

NS_ERROR_INTERNAL_ERRORRESULT_TYPEERROR

NS_ERROR_INVALID_ARG

NS_ERROR_INVALID_CONTENT_ENCODING

NS_ERROR_INVALID_POINTER

NS_ERROR_INVALID_SIGNATURE

NS_ERROR_IN_PROGRESS

NS_ERROR_LAUNCHED_CHILD_PROCESS

NS_ERROR_LOAD_SHOWED_ERRORPAGE

NS_ERROR_LOSS_OF_SIGNIFICANT_DATA

NS_ERROR_MALFORMED_URI

NS_ERROR_MALWARE_URI

NS_ERROR_MODULE_BASE

NS_ERROR_MODULE_BASE_OFFSET

NS_ERROR_MODULE_CALENDAR

NS_ERROR_MODULE_CONTENT

NS_ERROR_MODULE_DOM

NS_ERROR_MODULE_DOM_ANIM

NS_ERROR_MODULE_DOM_BLUETOOTH

NS_ERROR_MODULE_DOM_FILE

NS_ERROR_MODULE_DOM_FILEHANDLE

NS_ERROR_MODULE_DOM_FILESYSTEM

NS_ERROR_MODULE_DOM_INDEXEDDB

NS_ERROR_MODULE_DOM_MEDIA

NS_ERROR_MODULE_DOM_PUSH

NS_ERROR_MODULE_DOM_XPATH

NS_ERROR_MODULE_EDITOR

NS_ERROR_MODULE_ERRORRESULT

NS_ERROR_MODULE_FILES

NS_ERROR_MODULE_GENERAL

NS_ERROR_MODULE_GFX

NS_ERROR_MODULE_HTMLPARSER

NS_ERROR_MODULE_IMGLIB

NS_ERROR_MODULE_IPC

NS_ERROR_MODULE_LAYOUT

NS_ERROR_MODULE_LDAP

NS_ERROR_MODULE_MAILNEWS

NS_ERROR_MODULE_NETWORK

NS_ERROR_MODULE_PLUGINS

NS_ERROR_MODULE_PROFILE

NS_ERROR_MODULE_PYXPCOM

NS_ERROR_MODULE_RDF

NS_ERROR_MODULE_REG

NS_ERROR_MODULE_SCHEMA

NS_ERROR_MODULE_SECURITY

NS_ERROR_MODULE_SIGNED_APP

NS_ERROR_MODULE_SIGNED_JAR

NS_ERROR_MODULE_STORAGE

NS_ERROR_MODULE_SVG

NS_ERROR_MODULE_UCONV

NS_ERROR_MODULE_URILOADER

NS_ERROR_MODULE_URL_CLASSIFIER

NS_ERROR_MODULE_WIDGET

NS_ERROR_MODULE_XPCOM

NS_ERROR_MODULE_XPCONNECT

NS_ERROR_MODULE_XSLT

NS_ERROR_NET_INADEQUATE_SECURITY

NS_ERROR_NET_INTERRUPT

NS_ERROR_NET_PARTIAL_TRANSFER

NS_ERROR_NET_RESET

NS_ERROR_NET_TIMEOUT

NS_ERROR_NOT_AVAILABLE

NS_ERROR_NOT_CONNECTED

NS_ERROR_NOT_IMPLEMENTED

NS_ERROR_NOT_INITIALIZED

NS_ERROR_NOT_IN_TREE

NS_ERROR_NOT_RESUMABLE

NS_ERROR_NOT_SAME_THREAD

NS_ERROR_NO_AGGREGATION

NS_ERROR_NO_CONTENT

NS_ERROR_NO_INTERFACE

NS_ERROR_NULL_POINTER

NS_ERROR_OBJECT_IS_IMMUTABLE

NS_ERROR_OFFLINE

NS_ERROR_OUT_OF_MEMORY

NS_ERROR_PARSED_DATA_CACHED

NS_ERROR_PHISHING_URI

NS_ERROR_PLUGINS_PLUGINSNOTCHANGED

NS_ERROR_PLUGIN_BLOCKLISTED

NS_ERROR_PLUGIN_CLICKTOPLAY

NS_ERROR_PLUGIN_DISABLED

NS_ERROR_PLUGIN_TIME_RANGE_NOT_SUPPORTED

NS_ERROR_PORT_ACCESS_NOT_ALLOWED

NS_ERROR_PROXY_CONNECTION_REFUSED

NS_ERROR_RANGE_ERR

NS_ERROR_REDIRECT_LOOP

NS_ERROR_REMOTE_XUL

NS_ERROR_SAVE_LINK_AS_TIMEOUT

NS_ERROR_SERVICE_IN_USE

NS_ERROR_SERVICE_NOT_AVAILABLE

NS_ERROR_SERVICE_NOT_FOUND

NS_ERROR_SIGNED_APP_MANIFEST_INVALID

NS_ERROR_SIGNED_JAR_ENTRY_INVALID

NS_ERROR_SIGNED_JAR_ENTRY_MISSING

NS_ERROR_SIGNED_JAR_ENTRY_TOO_LARGE

NS_ERROR_SIGNED_JAR_MANIFEST_INVALID

NS_ERROR_SIGNED_JAR_MODIFIED_ENTRY

NS_ERROR_SIGNED_JAR_NOT_SIGNED

NS_ERROR_SIGNED_JAR_UNSIGNED_ENTRY

NS_ERROR_SIGNED_JAR_WRONG_SIGNATURE

NS_ERROR_SOCKET_ADDRESS_IN_USE

NS_ERROR_SOCKET_ADDRESS_NOT_SUPPORTED

NS_ERROR_SOCKET_CREATE_FAILED

NS_ERROR_SRI_CORRUPT

NS_ERROR_SRI_DISABLED

NS_ERROR_SRI_IMPORT

NS_ERROR_SRI_NOT_ELIGIBLE

NS_ERROR_SRI_UNEXPECTED_HASH_TYPE

NS_ERROR_STORAGE_BUSY

NS_ERROR_STORAGE_CONSTRAINT

NS_ERROR_STORAGE_IOERR

NS_ERROR_TRACKING_URI

NS_ERROR_TRANSPORT_INIT

NS_ERROR_TYPE_ERR

NS_ERROR_UCONV_NOCONV

NS_ERROR_UC_UPDATE_BUILD_PREFIX_FAILURE

NS_ERROR_UC_UPDATE_CHECKSUM_MISMATCH

NS_ERROR_UC_UPDATE_DUPLICATE_PREFIX

NS_ERROR_UC_UPDATE_FAIL_TO_WRITE_DISK

NS_ERROR_UC_UPDATE_INFINITE_LOOP

NS_ERROR_UC_UPDATE_MISSING_CHECKSUM

NS_ERROR_UC_UPDATE_PROTOCOL_PARSER_ERROR

NS_ERROR_UC_UPDATE_SHUTDOWNING

NS_ERROR_UC_UPDATE_TABLE_NOT_FOUND

NS_ERROR_UC_UPDATE_UNKNOWN

NS_ERROR_UC_UPDATE_WRONG_REMOVAL_INDICES

NS_ERROR_UDEC_ILLEGALINPUT

NS_ERROR_UENC_NOMAPPING

NS_ERROR_UNCATCHABLE_EXCEPTION

NS_ERROR_UNEXPECTED

NS_ERROR_UNKNOWN_HOST

NS_ERROR_UNKNOWN_PROTOCOL

NS_ERROR_UNKNOWN_PROXY_HOST

NS_ERROR_UNKNOWN_SOCKET_TYPE

NS_ERROR_UNORM_MOREOUTPUT

NS_ERROR_UNSAFE_CONTENT_TYPE

NS_ERROR_UNWANTED_URI

NS_ERROR_WONT_HANDLE_CONTENT

NS_ERROR_XBL_BLOCKED

NS_ERROR_XPATH_BAD_ARGUMENT_COUNT

NS_ERROR_XPATH_BAD_BANG

NS_ERROR_XPATH_BAD_COLON

NS_ERROR_XPATH_BAD_EXTENSION_FUNCTION

NS_ERROR_XPATH_BINARY_EXPECTED

NS_ERROR_XPATH_BRACKET_EXPECTED

NS_ERROR_XPATH_ILLEGAL_CHAR

NS_ERROR_XPATH_INVALID_ARG

NS_ERROR_XPATH_INVALID_AXIS

NS_ERROR_XPATH_INVALID_EXPRESSION_EVALUATED

NS_ERROR_XPATH_INVALID_VAR_NAME

NS_ERROR_XPATH_NO_NODE_TYPE_TEST

NS_ERROR_XPATH_OPERATOR_EXPECTED

NS_ERROR_XPATH_PAREN_EXPECTED

NS_ERROR_XPATH_PARSE_FAILURE

NS_ERROR_XPATH_UNBALANCED_CURLY_BRACE

NS_ERROR_XPATH_UNCLOSED_LITERAL

NS_ERROR_XPATH_UNEXPECTED_END

NS_ERROR_XPATH_UNKNOWN_FUNCTION

NS_ERROR_XPC_BAD_CID

NS_ERROR_XPC_BAD_CONVERT_JS

NS_ERROR_XPC_BAD_CONVERT_JS_NULL_REF

NS_ERROR_XPC_BAD_CONVERT_JS_ZERO_ISNOT_NULL

NS_ERROR_XPC_BAD_CONVERT_NATIVE

NS_ERROR_XPC_BAD_ID_STRING

NS_ERROR_XPC_BAD_IID

NS_ERROR_XPC_BAD_INITIALIZER_NAME

NS_ERROR_XPC_BAD_OP_ON_WN_PROTO

NS_ERROR_XPC_CALL_TO_SCRIPTABLE_FAILED

NS_ERROR_XPC_CANT_ADD_PROP_TO_WRAPPED_NATIVE

NS_ERROR_XPC_CANT_CALL_WO_SCRIPTABLE

NS_ERROR_XPC_CANT_CONVERT_OBJECT_TO_ARRAY

NS_ERROR_XPC_CANT_CONVERT_PRIMITIVE_TO_ARRAY

NS_ERROR_XPC_CANT_CONVERT_WN_TO_FUN

NS_ERROR_XPC_CANT_CREATE_WN

NS_ERROR_XPC_CANT_CTOR_WO_SCRIPTABLE

NS_ERROR_XPC_CANT_DEFINE_PROP_ON_WN

NS_ERROR_XPC_CANT_EXPORT_WN_STATIC

NS_ERROR_XPC_CANT_GET_ARRAY_INFO

NS_ERROR_XPC_CANT_GET_INTERFACE_INFO

NS_ERROR_XPC_CANT_GET_JSOBJECT_OF_DOM_OBJECT

NS_ERROR_XPC_CANT_GET_METHOD_INFO

NS_ERROR_XPC_CANT_GET_PARAM_IFACE_INFO

NS_ERROR_XPC_CANT_MODIFY_PROP_ON_WN

NS_ERROR_XPC_CANT_PASS_CPOW_TO_NATIVE

NS_ERROR_XPC_CANT_SET_OUT_VAL

NS_ERROR_XPC_CANT_SET_READ_ONLY_ATTRIBUTE

NS_ERROR_XPC_CANT_SET_READ_ONLY_CONSTANT

NS_ERROR_XPC_CANT_SET_READ_ONLY_METHOD

NS_ERROR_XPC_CANT_WATCH_WN_STATIC

NS_ERROR_XPC_CI_RETURNED_FAILURE

NS_ERROR_XPC_GS_RETURNED_FAILURE

NS_ERROR_XPC_HAS_BEEN_SHUTDOWN

NS_ERROR_XPC_INTERFACE_NOT_FROM_NSISUPPORTS

NS_ERROR_XPC_INTERFACE_NOT_SCRIPTABLE

NS_ERROR_XPC_JAVASCRIPT_ERROR

NS_ERROR_XPC_JAVASCRIPT_ERROR_WITH_DETAILS

NS_ERROR_XPC_JSOBJECT_HAS_NO_FUNCTION_NAMED

NS_ERROR_XPC_JS_THREW_EXCEPTION

NS_ERROR_XPC_JS_THREW_JS_OBJECT

NS_ERROR_XPC_JS_THREW_NATIVE_OBJECT

NS_ERROR_XPC_JS_THREW_NULL

NS_ERROR_XPC_JS_THREW_NUMBER

NS_ERROR_XPC_JS_THREW_STRING

NS_ERROR_XPC_NATIVE_RETURNED_FAILURE

NS_ERROR_XPC_NEED_OUT_OBJECT

NS_ERROR_XPC_NOT_ENOUGH_ARGS

NS_ERROR_XPC_NOT_ENOUGH_CHARS_IN_STRING

NS_ERROR_XPC_NOT_ENOUGH_ELEMENTS_IN_ARRAY

NS_ERROR_XPC_SCRIPTABLE_CALL_FAILED

NS_ERROR_XPC_SCRIPTABLE_CTOR_FAILED

NS_ERROR_XPC_SECURITY_MANAGER_VETO

NS_ERROR_XPC_UNEXPECTED

NS_ERROR_XSLT_ABORTED

NS_ERROR_XSLT_ALREADY_SET

NS_ERROR_XSLT_BAD_NODE_NAME

NS_ERROR_XSLT_BAD_RECURSION

NS_ERROR_XSLT_BAD_VALUE

NS_ERROR_XSLT_CALL_TO_KEY_NOT_ALLOWED

NS_ERROR_XSLT_EXECUTION_FAILURE

NS_ERROR_XSLT_LOAD_BLOCKED_ERROR

NS_ERROR_XSLT_LOAD_RECURSION

NS_ERROR_XSLT_NETWORK_ERROR

NS_ERROR_XSLT_NODESET_EXPECTED

NS_ERROR_XSLT_PARSE_FAILURE

NS_ERROR_XSLT_VAR_ALREADY_SET

NS_ERROR_XSLT_WRONG_MIME_TYPE

NS_EXACT_LENGTH

NS_FINDBROADCASTER_AWAIT_OVERLAYS

NS_FINDBROADCASTER_FOUND

NS_FINDBROADCASTER_NOT_FOUND

NS_HTMLPARSER_VALID_META_CHARSET

NS_HTMLTOKENS_NOT_AN_ENTITY

NS_HTML_STYLE_PROPERTY_NOT_THERE

NS_IMAGELIB_ERROR_FAILURE

NS_IMAGELIB_ERROR_NOT_FINISHED

NS_IMAGELIB_ERROR_NO_DECODER

NS_IMAGELIB_ERROR_NO_ENCODER

NS_IMAGELIB_SUCCESS_LOAD_FINISHED

NS_NET_STATUS_BEGIN_FTP_TRANSACTION

NS_NET_STATUS_CONNECTED_TO

NS_NET_STATUS_CONNECTING_TO

NS_NET_STATUS_END_FTP_TRANSACTION

NS_NET_STATUS_READING

NS_NET_STATUS_RECEIVING_FROM

NS_NET_STATUS_RESOLVED_HOST

NS_NET_STATUS_RESOLVING_HOST

NS_NET_STATUS_SENDING_TO

NS_NET_STATUS_TLS_HANDSHAKE_ENDED

NS_NET_STATUS_TLS_HANDSHAKE_STARTING

NS_NET_STATUS_WAITING_FOR

NS_NET_STATUS_WRITING

NS_NOINTERFACE

NS_OK

NS_OK_HAD_REPLACEMENTS

NS_OK_NO_NAME_CLAUSE_HANDLED

NS_OK_PARSE_SHEET

NS_OK_UDEC_EXACTLENGTH

NS_OK_UDEC_MOREINPUT

NS_OK_UDEC_MOREOUTPUT

NS_OK_UDEC_NOBOMFOUND

NS_OK_UENC_EXACTLENGTH

NS_OK_UENC_MOREINPUT

NS_OK_UENC_MOREOUTPUT

NS_PARTIAL_MORE_INPUT

NS_PARTIAL_MORE_OUTPUT

NS_POSITION_BEFORE_TABLE

NS_PROPTABLE_PROP_NOT_THERE

NS_PROPTABLE_PROP_OVERWRITTEN

NS_RDF_ASSERTION_ACCEPTED

NS_RDF_ASSERTION_REJECTED

NS_RDF_CURSOR_EMPTY

NS_RDF_NO_VALUE

NS_RDF_STOP_VISIT

NS_REFRESHURI_HEADER_FOUND

NS_STATE_PROPERTY_EXISTS

NS_STATE_PROPERTY_NOT_THERE

NS_SUCCESS_ADOPTED_DATA

NS_SUCCESS_AGGREGATE_RESULT

NS_SUCCESS_AUTH_FINISHED

NS_SUCCESS_DOM_NO_OPERATION

NS_SUCCESS_DOM_SCRIPT_EVALUATION_THREW

NS_SUCCESS_DOM_SCRIPT_EVALUATION_THREW_UNCATCHABLE

NS_SUCCESS_DONT_FIXUP

NS_SUCCESS_EDITOR_ELEMENT_NOT_FOUND

NS_SUCCESS_EDITOR_FOUND_TARGET

NS_SUCCESS_EVENT_CONSUMED

NS_SUCCESS_EVENT_HANDLED_ASYNCHRONOUSLY

NS_SUCCESS_FILE_DIRECTORY_EMPTY

NS_SUCCESS_INTERRUPTED_TRAVERSE

NS_SUCCESS_LOSS_OF_INSIGNIFICANT_DATA

NS_SUCCESS_RESTART_APP

NS_SUCCESS_RESTART_APP_NOT_SAME_PROFILE

NS_SUCCESS_UNORM_NOTFOUND

NS_SUCCESS_USING_FALLBACK_LOCALE

NS_TABLELAYOUT_CELL_NOT_FOUND

NS_XSLT_GET_NEW_HANDLER

Traits

NsresultExt
An extension trait which is intended to add methods to nsresult types. Unfortunately, due to ABI issues, this trait is implemented on all u32 types. These methods are meaningless on non-nsresult values.

Type Definitions

nsresult
The type of errors in gecko. This type is currently a type alias, rather than a newtype, in order to conform to the C ABI. In future versions of rust which support RFC #1758 or similar we may be able to use #[repr(transparent)] to get a better API for using nsresult.