Type Definition nserror::nsresult

type nsresult = u32;

The type of errors in gecko. This type is currently a type alias, rather than a newtype, in order to conform to the C ABI. In future versions of rust which support RFC #1758 or similar we may be able to use #[repr(transparent)] to get a better API for using nsresult.

The most unfortunate thing about this current implementation is that u32 and nsresult unify.